Remove "til og med"-date
A Grunnbeløp should always last until a new one is defined. We can't
have gaps, and having a TOM-date is prone for manually messing up.

The logic that picks the correct date is dependent that the elements are
sorted in the right order, so make an explicit sort of the Set, and not
depend on being defined in the right order.